+++ src/j1client.ts
@@ -5,12 +5,27 @@ import { InMemoryCache } from 'apollo-cache-inmemory';
 import { ApolloLink } from 'apollo-link';
 import { RetryLink } from 'apollo-link-retry';
 import { BatchHttpLink } from 'apollo-link-batch-http';
-import gql from 'graphql-tag';
 import fetch from 'node-fetch';
+
+import {
+  CREATE_ENTITY,
+  UPDATE_ENTITY,
+  DELETE_ENTITY,
+  CREATE_RELATIONSHIP,
+  UPSERT_ENTITY_RAW_DATA,
+  QUERY_V1,
+  CREATE_ALERT_RULE,
+  UPDATE_ALERT_RULE,
+  CREATE_QUESTION,
+  UPDATE_QUESTION,
+  DELETE_QUESTION,
+} from './queries';
 
 const J1_USER_POOL_ID_PROD = 'us-east-2_9fnMVHuxD';
 const J1_CLIENT_ID_PROD = '1hcv141pqth5f49df7o28ngq1u';
 const QUERY_RESULTS_TIMEOUT = 1000 * 60 * 5; // Poll s3 location for 5 minutes before timeout.
+const J1QL_SKIP_COUNT = 250;
+const J1QL_LIMIT_COUNT = 250;
 
 const JobStatus = {
   IN_PROGRESS: 'IN_PROGRESS',
@@ -169,7 +184,7 @@ export class JupiterOneClient {
     let results = [];
 
     while (!complete) {
-      let j1qlForPage = `${j1ql} SKIP ${
+      const j1qlForPage = `${j1ql} SKIP ${
         page * J1QL_SKIP_COUNT
       } LIMIT ${J1QL_LIMIT_COUNT}`;
 
@@ -188,7 +203,7 @@ export class JupiterOneClient {
       const deferredUrl = res.data.queryV1.url;
       let status = JobStatus.IN_PROGRESS;
       let statusFile;
-      let startTimeInMs = Date.now();
+      const startTimeInMs = Date.now();
       do {
         if (Date.now() - startTimeInMs > QUERY_RESULTS_TIMEOUT) {
           throw new Error(
@@ -462,259 +477,3 @@ export class JupiterOneClient {
     return res.data.deleteQuestion;
   }
 }
-
-const J1QL_SKIP_COUNT = 250;
-const J1QL_LIMIT_COUNT = 250;
-
-const CREATE_ENTITY = gql`
-  mutation CreateEntity(
-    $entityKey: String!
-    $entityType: String!
-    $entityClass: [String!]!
-    $properties: JSON
-  ) {
-    createEntity(
-      entityKey: $entityKey
-      entityType: $entityType
-      entityClass: $entityClass
-      properties: $properties
-    ) {
-      entity {
-        _id
-      }
-      vertex {
-        id
-        entity {
-          _id
-        }
-      }
-    }
-  }
-`;
-
-const UPDATE_ENTITY = gql`
-  mutation UpdateEntity($entityId: String!, $properties: JSON) {
-    updateEntity(entityId: $entityId, properties: $properties) {
-      entity {
-        _id
-      }
-      vertex {
-        id
-      }
-    }
-  }
-`;
-
-const DELETE_ENTITY = gql`
-  mutation DeleteEntity(
-    $entityId: String!
-    $timestamp: Long
-    $hardDelete: Boolean
-  ) {
-    deleteEntity(
-      entityId: $entityId
-      timestamp: $timestamp
-      hardDelete: $hardDelete
-    ) {
-      entity {
-        _id
-        _deleted
-        _endOn
-      }
-      vertex {
-        id
-        properties
-      }
-    }
-  }
-`;
-
-const CREATE_RELATIONSHIP = gql`
-  mutation CreateRelationship(
-    $relationshipKey: String!
-    $relationshipType: String!
-    $relationshipClass: String!
-    $fromEntityId: String!
-    $toEntityId: String!
-    $properties: JSON
-  ) {
-    createRelationship(
-      relationshipKey: $relationshipKey
-      relationshipType: $relationshipType
-      relationshipClass: $relationshipClass
-      fromEntityId: $fromEntityId
-      toEntityId: $toEntityId
-      properties: $properties
-    ) {
-      relationship {
-        _id
-      }
-      edge {
-        id
-        toVertexId
-        fromVertexId
-        relationship {
-          _id
-        }
-        properties
-      }
-    }
-  }
-`;
-
-const UPSERT_ENTITY_RAW_DATA = gql`
-  mutation UpsertEntityRawData(
-    $entityId: String!
-    $source: String!
-    $rawData: [JSON!]!
-  ) {
-    upsertEntityRawData(
-      entityId: $entityId
-      source: $source
-      rawData: $rawData
-    ) {
-      status
-    }
-  }
-`;
-
-const QUERY_V1 = gql`
-  query QueryLanguageV1(
-    $query: String!
-    $variables: JSON
-    $includeDeleted: Boolean
-    $deferredResponse: DeferredResponseOption
-    $deferredFormat: DeferredResponseFormat
-  ) {
-    queryV1(
-      query: $query
-      variables: $variables
-      includeDeleted: $includeDeleted
-      deferredResponse: $deferredResponse
-      deferredFormat: $deferredFormat
-    ) {
-      type
-      data
-      url
-    }
-  }
-`;
-
-const CREATE_ALERT_RULE = gql`
-  mutation CreateQuestionRuleInstance(
-    $instance: CreateQuestionRuleInstanceInput!
-  ) {
-    createQuestionRuleInstance(instance: $instance) {
-      id
-      name
-    }
-  }
-`;
-
-const UPDATE_ALERT_RULE = gql`
-  mutation UpdateQuestionRuleInstance(
-    $instance: UpdateQuestionRuleInstanceInput!
-  ) {
-    updateQuestionRuleInstance(instance: $instance) {
-      id
-      name
-      description
-      version
-      specVersion
-      latest
-      deleted
-      pollingInterval
-      templates
-      question {
-        queries {
-          name
-          query
-          version
-        }
-      }
-      operations {
-        when
-        actions
-      }
-      outputs
-    }
-  }
-`;
-
-const CREATE_QUESTION = gql`
-  mutation CreateQuestion($question: CreateQuestionInput!) {
-    createQuestion(question: $question) {
-      id
-      title
-      description
-      queries {
-        query
-        version
-      }
-      variables {
-        name
-        required
-        default
-      }
-      compliance {
-        standard
-        requirements
-      }
-      tags
-      accountId
-      integrationDefinitionId
-    }
-  }
-`;
-
-const UPDATE_QUESTION = gql`
-  mutation UpdateQuestion($id: ID!, $update: QuestionUpdate!) {
-    updateQuestion(id: $id, update: $update) {
-      id
-      title
-      description
-      queries {
-        query
-        version
-      }
-      variables {
-        name
-        required
-        default
-      }
-      compliance {
-        standard
-        requirements
-      }
-      tags
-      accountId
-      integrationDefinitionId
-    }
-  }
-`;
-
-const DELETE_QUESTION = gql`
-  mutation DeleteQuestion($id: ID!) {
-    deleteQuestion(id: $id) {
-      id
-      title
-      description
-      queries {
-        query
-        version
-      }
-      variables {
-        name
-        required
-        default
-      }
-      compliance {
-        standard
-        requirements
-      }
-      tags
-      accountId
-      integrationDefinitionId
-    }
-  }
-`;
