+++ src/util/error.ts
@@ -1,6 +1,6 @@
 import chalk from 'chalk';
 
-export function fatal(message: string, code: number = 1) {
+export function fatal(message: string, code = 1) {
   console.log(chalk.red(message));
   process.exit(code);
 }
